Chess Game Initial Commit
This is the first commit and contains all the different type of piece
classes( Horse, King, Queen, Pawn, Bishop, Rook),  the Piece class, Game
class, Player Class, ChessGUI, and Spot class. We have just finished
creating all the piece classes, creating the game and spot class, and
using spot object as a button for the gui.